backend_src/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "falcon-backend"
version = "0.1.0"
description = "FastAPI + SQLAlchemy backend with MCP integration"
authors = [
    { name = "Nicolas Janis", email = "nicolas.d.janis@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic[email]",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "websockets>=11.0.0",
    "loguru>=0.7.0",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-google-genai",
    "langchain-mcp-adapters",
    "openai",
    "ollama",
    "watchdog",
    "sqlalchemy",
    "psycopg2-binary",
    "python-jose",
    "pyjwt",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2"
]

[tool.setuptools.packages.find]
where = ["."]
